+++ packages/ui/src/components/widgets/SwapWidget/index.tsx
@@ -4,7 +4,6 @@ import { useRelayClient } from '../../../hooks/index.js'
 import type { Address } from 'viem'
 import { formatUnits } from 'viem'
 import type { LinkedWallet, Token } from '../../../types/index.js'
-import { AnchorButton } from '../../primitives/Anchor.js'
 import { formatFixedLength, formatDollar } from '../../../utils/numbers.js'
 import AmountInput from '../../common/AmountInput.js'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
@@ -53,6 +52,7 @@ type BaseSwapWidgetProps = {
   supportedWalletVMs: ChainVM[]
   disableInputAutoFocus?: boolean
   popularChainIds?: number[]
+  disablePasteWalletAddressOption?: boolean
   onFromTokenChange?: (token?: Token) => void
   onToTokenChange?: (token?: Token) => void
   onConnectWallet?: () => void
@@ -100,6 +100,7 @@ const SwapWidget: FC<SwapWidgetProps> = ({
   supportedWalletVMs,
   disableInputAutoFocus = false,
   popularChainIds,
+  disablePasteWalletAddressOption,
   onSetPrimaryWallet,
   onLinkNewWallet,
   onFromTokenChange,
@@ -109,7 +110,7 @@ const SwapWidget: FC<SwapWidgetProps> = ({
   onSwapSuccess,
   onSwapValidating,
   onSwapError
-}) => {
+}): JSX.Element => {
   const relayClient = useRelayClient()
   const providerOptionsContext = useContext(ProviderOptionsContext)
   const connectorKeyOverrides = providerOptionsContext.vmConnectorKeyOverrides
@@ -334,6 +335,13 @@ const SwapWidget: FC<SwapWidgetProps> = ({
           connectorKeyOverrides
         ])
 
+        //Handle if the paste wallet address option is disabled while there is a custom to address
+        useEffect(() => {
+          if (disablePasteWalletAddressOption && customToAddress) {
+            setCustomToAddress(undefined)
+          }
+        }, [disablePasteWalletAddressOption])
+
         const promptSwitchRoute =
           (isCapacityExceededError || isCouldNotExecuteError) &&
           supportsExternalLiquidity &&
@@ -429,6 +437,9 @@ const SwapWidget: FC<SwapWidgetProps> = ({
                           <MultiWalletDropdown
                             context="origin"
                             selectedWalletAddress={address}
+                            disablePasteWalletAddressOption={
+                              disablePasteWalletAddressOption
+                            }
                             onSelect={(wallet) =>
                               onSetPrimaryWallet?.(wallet.address)
                             }
@@ -463,8 +474,8 @@ const SwapWidget: FC<SwapWidgetProps> = ({
                             tradeType === 'EXACT_INPUT'
                               ? amountInputValue
                               : amountInputValue
-                                ? formatFixedLength(amountInputValue, 8)
-                                : amountInputValue
+                              ? formatFixedLength(amountInputValue, 8)
+                              : amountInputValue
                           }
                           setValue={(e) => {
                             setAmountInputValue(e)
@@ -528,13 +539,13 @@ const SwapWidget: FC<SwapWidgetProps> = ({
                             isSingleChainLocked
                               ? [lockChainId]
                               : isChainLocked(
-                                    fromToken?.chainId,
-                                    lockChainId,
-                                    toToken?.chainId,
-                                    lockFromToken
-                                  ) && fromToken?.chainId
-                                ? [fromToken.chainId]
-                                : undefined
+                                  fromToken?.chainId,
+                                  lockChainId,
+                                  toToken?.chainId,
+                                  lockFromToken
+                                ) && fromToken?.chainId
+                              ? [fromToken.chainId]
+                              : undefined
                           }
                           chainIdsFilter={
                             !fromChainWalletVMSupported && toToken
@@ -752,6 +763,9 @@ const SwapWidget: FC<SwapWidgetProps> = ({
                         toChainWalletVMSupported ? (
                           <MultiWalletDropdown
                             context="destination"
+                            disablePasteWalletAddressOption={
+                              disablePasteWalletAddressOption
+                            }
                             selectedWalletAddress={recipient}
                             onSelect={(wallet) =>
                               setCustomToAddress(wallet.address)
@@ -840,8 +854,8 @@ const SwapWidget: FC<SwapWidgetProps> = ({
                             tradeType === 'EXPECTED_OUTPUT'
                               ? amountOutputValue
                               : amountOutputValue
-                                ? formatFixedLength(amountOutputValue, 8)
-                                : amountOutputValue
+                              ? formatFixedLength(amountOutputValue, 8)
+                              : amountOutputValue
                           }
                           setValue={(e) => {
                             setAmountOutputValue(e)
@@ -920,13 +934,13 @@ const SwapWidget: FC<SwapWidgetProps> = ({
                             isSingleChainLocked
                               ? [lockChainId]
                               : isChainLocked(
-                                    toToken?.chainId,
-                                    lockChainId,
-                                    fromToken?.chainId,
-                                    lockToToken
-                                  ) && toToken?.chainId
-                                ? [toToken.chainId]
-                                : undefined
+                                  toToken?.chainId,
+                                  lockChainId,
+                                  fromToken?.chainId,
+                                  lockToToken
+                                ) && toToken?.chainId
+                              ? [toToken.chainId]
+                              : undefined
                           }
                           chainIdsFilter={
                             !fromChainWalletVMSupported && fromToken