+++ packages/relay-bitcoin-wallet-adapter/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @reservoir0x/relay-bitcoin-wallet-adapter
 
+## 3.0.12
+
+### Patch Changes
+
+- Updated dependencies [66b8de8]
+  - @reservoir0x/relay-sdk@1.6.12
+
+## 3.0.11
+
+### Patch Changes
+
+- Updated dependencies [0af5808]
+  - @reservoir0x/relay-sdk@1.6.11
+
 ## 3.0.10
 
 ### Patch Changes